+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2020-04-20
+- added FrontendAsset service
+- fixed assets not loading in contao 4.9
+
 ## [1.1.0] - 2020-03-19
 - updated symfony dependencies to reflect the actual bundle usage und support contao 4.9 (thanks to fritzmg, #3, #4)
 - update encore bundle config, required now minimum version 1.5